nfc(major-refactor): made Layer a base class and added Dense layer
