+++ main.cpp
@@ -7,9 +7,9 @@ int main(int argc, char *argv[])
     Network network;
     std::shared_ptr<Optimizer> AdamOptimizer = std::make_shared<Adam>(1);
 
-    std::shared_ptr<Layer> layer1 = std::make_shared<Layer>(3, ACTIVATION::SIGMOID, WEIGHT_INIT::GLOROT);
-    std::shared_ptr<Layer> layer2 = std::make_shared<Layer>(2, ACTIVATION::SIGMOID, WEIGHT_INIT::HE);
-    std::shared_ptr<Layer> layerOuput = std::make_shared<Layer>(2, ACTIVATION::SIGMOID, WEIGHT_INIT::GLOROT);
+    std::shared_ptr<Layer> layer1 = std::make_shared<Dense>(3, ACTIVATION::SIGMOID, WEIGHT_INIT::GLOROT);
+    std::shared_ptr<Layer> layer2 = std::make_shared<Dense>(2, ACTIVATION::SIGMOID, WEIGHT_INIT::HE);
+    std::shared_ptr<Layer> layerOuput = std::make_shared<Dense>(2, ACTIVATION::SIGMOID, WEIGHT_INIT::GLOROT);
 
     network.addLayer(layer1);
     network.addLayer(layer2);
@@ -26,17 +26,17 @@ int main(int argc, char *argv[])
     std::cout << "num of layers : " << network.getNumLayers() << "\n";
 
     std::cout
-        << "Input Layer before training : "
+        << "Input Dense before training : "
         << "\n";
     layer1->printWeights();
     layer1->printOutputs();
 
-    std::cout << "Layer 2 before training : "
+    std::cout << "Dense 2 before training : "
               << "\n";
     layer2->printWeights();
     layer2->printOutputs();
 
-    std::cout << "Output Layer before training : "
+    std::cout << "Output Dense before training : "
               << "\n";
     layerOuput->printWeights();
     layerOuput->printOutputs();
@@ -51,17 +51,17 @@ int main(int argc, char *argv[])
     std::shared_ptr<Layer> test = network.getLayer(1);
     std::shared_ptr<Layer> test2 = network.getLayer(2);
 
-    std::cout << "Input Layer after training : "
+    std::cout << "Input Dense after training : "
               << "\n";
     input->printWeights();
     input->printOutputs();
 
-    std::cout << "Layer 2 after training : "
+    std::cout << "Dense 2 after training : "
               << "\n";
     test->printWeights();
     test->printOutputs();
 
-    std::cout << "Output Layer after training : "
+    std::cout << "Output Dense after training : "
               << "\n";
     test2->printWeights();
     test2->printOutputs();
