+++ src/CMakeLists.txt
@@ -2,7 +2,6 @@
 set(SRC_FILES 
   ${NETWORK_DIR}/Flatten.cpp
   ${NETWORK_DIR}/Network.cpp
-  ${NETWORK_DIR}/Layer.cpp
 )
 
 add_library(NeuralNet ${SRC_FILES})
