--- src/NeuralNet/Layer.cpp
@@ -1,188 +1,0 @@
-#include "Layer.hpp"
-
-using namespace NeuralNet;
-
-Layer::Layer(int nNeurons, ACTIVATION activation, WEIGHT_INIT weightInit, int bias)
-{
-    this->bias = bias;
-    this->nNeurons = nNeurons;
-    this->weightInit = weightInit;
-    this->activation = activation;
-    this->setActivation(activation);
-}
-
-int Layer::getNumNeurons() const
-{
-    return nNeurons;
-}
-
-Eigen::MatrixXd Layer::getOutputs() const
-{
-    return outputs;
-}
-
-Eigen::MatrixXd Layer::getWeights() const
-{
-    return weights;
-}
-
-void Layer::printWeights()
-{
-    std::cout << this->weights << "\n";
-    return;
-}
-
-void Layer::printOutputs()
-{
-    std::cout << this->outputs << "\n";
-    return;
-}
-
-/**
- * PRIVATE METHODS
- */
-/**
- * This function will be used to properly initialize the Layer
- * It's being done like this because of the number of neurons of the previous layer that's unkown prior
- */
-void Layer::init(int numRows)
-{
-    // First and foremost init the biases and the outputs
-    double mean = 0, stddev = 0;
-    this->weights = Eigen::MatrixXd ::Zero(numRows, nNeurons);
-
-    // calculate mean and stddev based on init algo
-    switch (this->weightInit)
-    {
-    case WEIGHT_INIT::GLOROT:
-        // sqrt(fan_avg)
-        stddev = sqrt(static_cast<double>(2) / (numRows + nNeurons));
-        break;
-    case WEIGHT_INIT::HE:
-        // sqrt(2/fan_in)
-        stddev = sqrt(static_cast<double>(2) / numRows);
-        break;
-    case WEIGHT_INIT::LECUN:
-        // sqrt(1/fan_in)
-        stddev = sqrt(static_cast<double>(1) / numRows);
-        break;
-    default:
-        break;
-    }
-
-    // Init the weights
-    this->weightInit == WEIGHT_INIT::RANDOM ? randomWeightInit(&(this->weights), -1, 1) : randomDistWeightInit(&(this->weights), mean, stddev);
-}
-
-void Layer::setActivation(ACTIVATION activation)
-{
-    switch (activation)
-    {
-    case ACTIVATION::SIGMOID:
-        this->activate = Sigmoid::activate;
-        this->diff = Sigmoid::diff;
-        break;
-    case ACTIVATION::RELU:
-        this->activate = Relu::activate;
-        this->diff = Relu::diff;
-        break;
-    case ACTIVATION::SOFTMAX:
-        this->activate = Softmax::activate;
-        this->diff = Softmax::diff;
-        break;
-    /**
-     * Add cases as I add activations
-     */
-    default:
-        assert(false && "Activation not defined");
-    }
-
-    return;
-}
-
-void Layer::feedInputs(std::vector<double> inputs)
-{
-    this->feedInputs(Eigen::MatrixXd::Map(&inputs[0], inputs.size(), 1));
-    return;
-}
-
-void Layer::feedInputs(std::vector<std::vector<std::vector<double>>> inputs)
-{
-    assert(false && "Cannot feed 3d vectors, a Flatten layer could do it though");
-    return;
-}
-
-void Layer::feedInputs(Eigen::MatrixXd inputs)
-{
-    // If the layer is an "input" layer
-    if (weights.rows() == 0)
-    {
-        this->setOutputs(inputs);
-        return;
-    }
-    inputs = inputs.cols() == weights.rows() ? inputs : inputs.transpose();
-
-    assert(inputs.cols() == weights.rows());
-    this->computeOutputs(inputs);
-    return;
-}
-
-void Layer::computeOutputs(Eigen::MatrixXd inputs)
-{
-    // Initialize the biases based on the input's size
-    if (biases.rows() == 0 && biases.cols() == 0)
-    {
-        biases = Eigen::MatrixXd::Constant(inputs.rows(), nNeurons, bias);
-    }
-
-    // Weighted sum
-    Eigen::MatrixXd wSum = inputs * weights;
-
-    wSum = wSum.array() + biases.array();
-
-    outputs = activate(wSum);
-    return;
-}
-
-void Layer::setOutputs(std::vector<double> outputs)
-{
-    assert(outputs.size() == nNeurons);
-    this->outputs = Eigen::MatrixXd ::Map(&outputs[0], this->getNumNeurons(), 1);
-}
-
-void Layer::setOutputs(Eigen::MatrixXd outputs)
-{
-    this->outputs = outputs;
-}
-
-void Layer::randomWeightInit(Eigen::MatrixXd *weights, double min, double max)
-{
-    for (int col = 0; col < weights->cols(); col++)
-    {
-        for (int row = 0; row < weights->rows(); row++)
-        {
-            weights->operator()(row, col) = mtRand(min, max);
-        }
-    }
-
-    return;
-}
-
-void Layer::randomDistWeightInit(Eigen::MatrixXd *weights, double mean, double stddev)
-{
-    std::random_device rseed;
-    std::default_random_engine generator(rseed());
-    std::normal_distribution<double> distribution(mean, stddev);
-
-    for (int col = 0; col < weights->cols(); col++)
-    {
-        for (int row = 0; row < weights->rows(); row++)
-        {
-            weights->operator()(row, col) = distribution(generator);
-        }
-    }
-
-    return;
-}
-
-Layer::~Layer() {}
